Develop

+++ module/OAuth/App/service-rest/service/src/main/java/org/jepria/oauth/token/TokenServiceImpl.java
@@ -5,14 +5,6 @@ import org.jepria.oauth.exception.OAuthRuntimeException;
 import org.jepria.oauth.key.KeyService;
 import org.jepria.oauth.key.dto.KeyDto;
 import org.jepria.oauth.sdk.GrantType;
-import org.jepria.oauth.session.SessionService;
-import org.jepria.oauth.session.dto.SessionCreateDto;
-import org.jepria.oauth.session.dto.SessionDto;
-import org.jepria.oauth.session.dto.SessionSearchDto;
-import org.jepria.oauth.session.dto.SessionUpdateDto;
-import org.jepria.oauth.token.TokenService;
-import org.jepria.oauth.token.dto.TokenDto;
-import org.jepria.oauth.token.dto.TokenInfoDto;
 import org.jepria.oauth.sdk.ResponseType;
 import org.jepria.oauth.sdk.token.Signer;
 import org.jepria.oauth.sdk.token.Token;
@@ -20,11 +12,16 @@ import org.jepria.oauth.sdk.token.TokenImpl;
 import org.jepria.oauth.sdk.token.Verifier;
 import org.jepria.oauth.sdk.token.rsa.SignerRSA;
 import org.jepria.oauth.sdk.token.rsa.VerifierRSA;
+import org.jepria.oauth.session.SessionService;
+import org.jepria.oauth.session.dto.SessionCreateDto;
+import org.jepria.oauth.session.dto.SessionDto;
+import org.jepria.oauth.session.dto.SessionSearchDto;
+import org.jepria.oauth.session.dto.SessionUpdateDto;
+import org.jepria.oauth.token.dto.TokenDto;
+import org.jepria.oauth.token.dto.TokenInfoDto;
 import org.jepria.server.service.security.Credential;
 
 import java.net.URI;
-import java.security.MessageDigest;
-import java.security.SecureRandom;
 import java.text.ParseException;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
@@ -33,8 +30,6 @@ import static org.jepria.oauth.sdk.OAuthConstants.*;
 
 public class TokenServiceImpl implements TokenService {
 
-  private static final int ACCESS_TOKEN_LIFETIME = 3600 * 8; //8 hours
-  private static final int REFRESH_TOKEN_LIFETIME = 3600 * 24 * 7; //1 week
   private static final String TOKEN_TYPE = "Bearer";
   private final SessionService sessionService;
   private final KeyService keyService;
@@ -68,7 +63,6 @@ public class TokenServiceImpl implements TokenService {
                                 String redirectUri,
                                 String accessTokenId,
                                 String refreshTokenId,
-                                Boolean hasToken,
                                 Credential credential) {
     SessionSearchDto searchTemplate = new SessionSearchDto();
     searchTemplate.setAuthorizationCode(authCode);
@@ -76,12 +70,11 @@ public class TokenServiceImpl implements TokenService {
     searchTemplate.setRedirectUri(redirectUri);
     searchTemplate.setAccessTokenId(accessTokenId);
     searchTemplate.setRefreshTokenId(refreshTokenId);
-    searchTemplate.setHasToken(hasToken);
     List<SessionDto> result = sessionService.find(searchTemplate, credential);
     if (result.size() == 1) {
       return result.get(0);
     } else {
-      throw new OAuthRuntimeException(INVALID_GRANT, "Authorization request not found");
+      return null;
     }
   }
 
@@ -111,7 +104,8 @@ public class TokenServiceImpl implements TokenService {
                          String issuer,
                          String authCode,
                          String clientId,
-                         URI redirectUri) {
+                         URI redirectUri,
+                         Integer accessTokenLifeTime) {
     if (ResponseType.TOKEN.equals(responseType)) {
       KeyDto keyDto = keyService.getKeys(null, serverCredential);
       if (authCode == null) {
@@ -128,18 +122,18 @@ public class TokenServiceImpl implements TokenService {
           redirectUri.toString(),
           null,
           null,
-          false,
           serverCredential);
+      if (session == null) throw new OAuthRuntimeException(INVALID_GRANT, "Authorization request not found");
       if (TimeUnit.MILLISECONDS.toMinutes(new Date().getTime() - session.getDateIns().getTime()) > 10) {
         throw new OAuthRuntimeException(INVALID_GRANT, "Authorization code active time has expired.");
       }
       if (session.getAccessTokenId() != null) {
         throw new OAuthRuntimeException(INVALID_GRANT, "Request is finished.");
       }
-      Token accessToken = generateToken(session.getOperatorLogin(), null, session.getOperator().getValue(), issuer, keyDto.getPrivateKey(), 4);
+      Token accessToken = generateToken(session.getOperatorLogin(), null, session.getOperator().getValue(), issuer, keyDto.getPrivateKey(), accessTokenLifeTime);
       TokenDto tokenDto = new TokenDto();
       tokenDto.setAccessToken(accessToken.asString());
-      tokenDto.setExpiresIn(ACCESS_TOKEN_LIFETIME);
+      tokenDto.setExpiresIn(new Long(accessTokenLifeTime * 3600));
       tokenDto.setTokenType(TOKEN_TYPE);
       updateSession(session,
           accessToken.getJti(),
@@ -156,7 +150,8 @@ public class TokenServiceImpl implements TokenService {
   public TokenDto create(String clientId,
                          String authCode,
                          String issuer,
-                         URI redirectUri) {
+                         URI redirectUri,
+                         Integer accessTokenLifeTime) {
     KeyDto keyDto = keyService.getKeys(null, serverCredential);
     if (authCode == null) {
       throw new OAuthRuntimeException(INVALID_REQUEST, "Authorization code is null.");
@@ -172,18 +167,18 @@ public class TokenServiceImpl implements TokenService {
         redirectUri.toString(),
         null,
         null,
-        false,
         serverCredential);
+    if (session == null) throw new OAuthRuntimeException(INVALID_GRANT, "Authorization request not found");
     if (TimeUnit.MILLISECONDS.toMinutes(new Date().getTime() - session.getDateIns().getTime()) > 10) {
       throw new OAuthRuntimeException(INVALID_GRANT, "Authorization code active time has expired.");
     }
     if (session.getAccessTokenId() != null) {
       throw new OAuthRuntimeException(INVALID_GRANT, "Request is finished.");
     }
-    Token accessToken = generateToken(session.getOperatorLogin(), null, session.getOperator().getValue(), issuer, keyDto.getPrivateKey(), 4);
+    Token accessToken = generateToken(session.getOperatorLogin(), null, session.getOperator().getValue(), issuer, keyDto.getPrivateKey(), accessTokenLifeTime);
     TokenDto tokenDto = new TokenDto();
     tokenDto.setAccessToken(accessToken.asString());
-    tokenDto.setExpiresIn(ACCESS_TOKEN_LIFETIME);
+    tokenDto.setExpiresIn(new Long(accessTokenLifeTime * 3600));
     tokenDto.setTokenType(TOKEN_TYPE);
     updateSession(session,
       accessToken.getJti(),
@@ -192,28 +187,33 @@ public class TokenServiceImpl implements TokenService {
       serverCredential);
     return tokenDto;
   }
-  
+
+  private void checkClientGrantTypes(String clientId, String grantType) {
+    List<String> clientGrantTypes = clientService.getClientGrantTypes(clientId);
+    if (clientGrantTypes.size() == 0 || !clientGrantTypes.stream().anyMatch(clientGrantType -> clientGrantType.equals(grantType))) {
+      throw new OAuthRuntimeException(UNAUTHORIZED_CLIENT, "Client doesn't have enough permissions to use responseType=" + grantType);
+    }
+  }
+
   @Override
   public TokenDto create(String clientId,
                          String username,
                          Integer userId,
-                         String issuer) {
-    List<String> clientGrantTypes = clientService.getClientGrantTypes(clientId);
-    if (clientGrantTypes.size() == 0 || !clientGrantTypes.stream().anyMatch(clientGrantType -> clientGrantType.equals(GrantType.PASSWORD))) {
-      throw new OAuthRuntimeException(UNAUTHORIZED_CLIENT, "Client doesn't have enough permissions to use responseType=" + GrantType.PASSWORD);
-    }
+                         String issuer,
+                         Integer accessTokenLifeTime,
+                         Integer refreshTokenLifeTime) {
+    checkClientGrantTypes(clientId, GrantType.PASSWORD);
     KeyDto keyDto = keyService.getKeys(null, serverCredential);
-    return createTokenPair(keyDto.getPrivateKey(), issuer, clientId, username, userId);
+    return createTokenPair(keyDto.getPrivateKey(), issuer, clientId, username, userId, accessTokenLifeTime, refreshTokenLifeTime);
   }
   
   @Override
   public TokenDto create(String clientId,
                          String refreshTokenString,
-                         String issuer) {
-    List<String> clientGrantTypes = clientService.getClientGrantTypes(clientId);
-    if (clientGrantTypes.size() == 0 || !clientGrantTypes.stream().anyMatch(clientGrantType -> clientGrantType.equals(GrantType.REFRESH_TOKEN))) {
-      throw new OAuthRuntimeException(UNAUTHORIZED_CLIENT, "Client doesn't have enough permissions to use responseType=" + GrantType.REFRESH_TOKEN);
-    }
+                         String issuer,
+                         Integer accessTokenLifeTime,
+                         Integer refreshTokenLifeTime) {
+    checkClientGrantTypes(clientId, GrantType.REFRESH_TOKEN);
     KeyDto keyDto = keyService.getKeys(null, serverCredential);
     try {
       Token refreshToken = TokenImpl.parseFromString(refreshTokenString);
@@ -224,11 +224,18 @@ public class TokenServiceImpl implements TokenService {
             null,
             null,
             refreshToken.getJti(),
-            true,
             serverCredential);
+        if (sessionDto == null) throw new OAuthRuntimeException(INVALID_GRANT, "Authorization request not found");
         String[] subject = refreshToken.getSubject().split(":");
         sessionService.deleteRecord(String.valueOf(sessionDto.getSessionId()), serverCredential);
-        return createTokenPair(keyDto.getPrivateKey(), issuer, refreshToken.getAudience().isEmpty() ? null : refreshToken.getAudience().get(0), subject[0], Integer.valueOf(subject[1]));
+        return createTokenPair(
+            keyDto.getPrivateKey(),
+            issuer,
+            refreshToken.getAudience().isEmpty() ? null : refreshToken.getAudience().get(0),
+            subject[0],
+            Integer.valueOf(subject[1]),
+            accessTokenLifeTime,
+            refreshTokenLifeTime);
       } else {
         throw new OAuthRuntimeException(INVALID_GRANT, "Token verification failed");
       }
@@ -240,25 +247,26 @@ public class TokenServiceImpl implements TokenService {
   @Override
   public TokenDto create(String clientId,
                          Integer clientOperatorId,
-                         String issuer) {
-    List<String> clientGrantTypes = clientService.getClientGrantTypes(clientId);
-    if (clientOperatorId == null || clientGrantTypes.size() == 0 || !clientGrantTypes.stream().anyMatch(clientGrantType -> clientGrantType.equals(GrantType.CLIENT_CREDENTIALS))) {
-      throw new OAuthRuntimeException(UNAUTHORIZED_CLIENT, "Client doesn't have enough permissions to use responseType=" + GrantType.CLIENT_CREDENTIALS);
-    }
+                         String issuer,
+                         Integer accessTokenLifeTime,
+                         Integer refreshTokenLifeTime) {
+    checkClientGrantTypes(clientId, GrantType.CLIENT_CREDENTIALS);
     KeyDto keyDto = keyService.getKeys(null, serverCredential);
-    return createTokenPair(keyDto.getPrivateKey(), issuer, clientId, clientId, clientOperatorId);
+    return createTokenPair(keyDto.getPrivateKey(), issuer, clientId, clientId, clientOperatorId, accessTokenLifeTime, refreshTokenLifeTime);
   }
   
   private TokenDto createTokenPair(String privateKeyString,
                                    String issuer,
                                    String clientId,
                                    String username,
-                                   Integer userId) {
-    Token accessToken = generateToken(username, Collections.singletonList(clientId), userId, issuer, privateKeyString, 1);
-    Token refreshToken = generateToken(username, Collections.singletonList(clientId), userId, issuer, privateKeyString, 24 * 7);
+                                   Integer userId,
+                                   Integer accessTokenLifeHours,
+                                   Integer refreshTokenLifeHours) {
+    Token accessToken = generateToken(username, Collections.singletonList(clientId), userId, issuer, privateKeyString, accessTokenLifeHours);
+    Token refreshToken = generateToken(username, Collections.singletonList(clientId), userId, issuer, privateKeyString, refreshTokenLifeHours);
     TokenDto tokenDto = new TokenDto();
     tokenDto.setAccessToken(accessToken.asString());
-    tokenDto.setExpiresIn(ACCESS_TOKEN_LIFETIME); // 1 hours active time
+    tokenDto.setExpiresIn(new Long(accessTokenLifeHours * 3600));
     tokenDto.setTokenType("Bearer");
     tokenDto.setRefreshToken(refreshToken.asString());
 
@@ -288,7 +296,6 @@ public class TokenServiceImpl implements TokenService {
           null,
           token.getJti(),
           null,
-          true,
           serverCredential);
       if (sessionDto == null) {
         result.setActive(false);
@@ -322,11 +329,15 @@ public class TokenServiceImpl implements TokenService {
           null,
           token.getJti(),
           null,
-          true,
           serverCredential);
-      sessionService.deleteRecord(String.valueOf(sessionDto.getSessionId()), serverCredential);
+      if (sessionDto != null) {
+        sessionService.deleteRecord(String.valueOf(sessionDto.getSessionId()), serverCredential);
+      }
     } catch (ParseException e) {
       throw new OAuthRuntimeException(SERVER_ERROR, e);
+    } catch (IllegalArgumentException ex) {
+      System.out.println("Failed token value: " + tokenString);
+      throw new OAuthRuntimeException(SERVER_ERROR, ex);
     }
   }
